+++ ChangeLog.rst
@@ -6,8 +6,7 @@ Note worthy changes
 
 - Django 1.10 compatibility.
 
-- The Twitter provider now supports querying of the email address.
-
+- The Twitter and GitHub providers now support querying of the email address.
 
 
 0.26.1 (2016-07-25)